Moving util/math.h -> util/cutlass_math.h to match HEAD of master. This is the simple solution to the use of multiple Cutlass versions in cuml (one for ml-prims and one for cuml algorithms).
